Maria Ramos:  Now I have added code into Import to take into account the possibility that the user's .csv file has: incorrect column names (column names that don't match with the column names in the Parse table), duplicate columns (more than one column with the same column name), and/or it is missing either a serial_number column or a room column.
My next step is to create sample csv files to easily test these scenarios, so to not have to keep modifying one file randomly. Also, I will attempt to write code to check if any serial numbers in the .csv file are already within the Parse database and, if so, then give the user a warning telling them that the serial numbers at specific rows are already in the database and give them the option to cancel so they may change the data in the .csv manually and try again or continue so to update the old values in the database to the new ones.